Using the rmsprop optimizer -0.46 dice loss is achieved
The training was not started from randomly initialised weights but was the continuation of the previous learning, which only differed in the use of the adam optimizer. This is why it seemed like that by using the rmsprop optimizer the training was significantly faster and better dice loss was achieved. Learning rate reduction on plateu was enabled during training which could prevent any significant improvement at the and of the training.